+++ solution/backend/resources/models/categories.py
