+++ solution/backend/resources/models/public_resources.py
@@ -1,7 +1,7 @@
 from django.db import models
 from django_jsonform.models.fields import ArrayField
 
-from .resource import AbstractPublicResource
+from .resources import AbstractPublicResource
 
 FR_ACTION_TYPES = [("RFI", "RFI"), ("NPRM", "NPRM"), ("Final", "Final")]
