+++ solution/backend/resources/old_views/categories.py
