+++ solution/backend/resources/old_views/locations.py
