+++ solution/backend/resources/old_views/mixins.py
