+++ solution/backend/resources/old_views/resources.py
