--- solution/backend/resources/templates/admin/import_resources.html
@@ -1,36 +1,0 @@
-{% extends "admin/base_site.html" %}
-
-{% block content %}
-<form action="." method="GET" enctype="multipart/form-data">
-    {% csrf_token %}
-    <p>The CSV file is required to have the following fields in order:</p>
-    <table>
-        <th>Name</th>
-        <th>Description</th>
-        <th>Url</th>
-        <th>Category</th>
-        <th>locations</th>
-    </table><br>
-    <p>You can get a copy of the template by clicking "Get Template". </p>
-    <p><input type="submit" name="get_template" value="Get Template"/>
-    <p> CSV's seperate fields by commas.  If you use "Google Sheets" or "Microsoft Excel" to make the csv you do not need to worry but if you use a text editor have the format be like below:</p>
-        <code>
-    Name, Description, URL, Category, Locations <br>
-    content-1, description-1, url-1, category-1, "location-1, location-2, location-3"<br><br>
-    </code>
-    Locations need to be in the following format: <br>
-    <ul>
-        <li>TitleNumber CFR PartNumber SectionNumber or subpart(ex 42 CFR 400 200, 42 CFR 400 subpart b) </li>
-        <li>TitleNumber PartNumber.SectionNumber(ex 42 400.200)</li>
-        <li>TitleNumber PartNumber SectionNumber or Subpart(ex 42 400 200, 42 400 subpart A)
-    </ul>
-
-</form>
-    <form action="." method="post" enctype="multipart/form-data">
-        {{ form.as_p }}
-        {% csrf_token %}
-        <p> Upload the content by using import resources</p><br>
-        <p> After submission you will be redirected to a new page to show the results of the import as well as any errors </p>
-        <input type="submit" name="import" value="Import Resources"/></p>
-    </form>
-{% endblock %}
