--- solution/backend/resources/templates/admin/import_resources_button.html
@@ -1,23 +1,0 @@
-{% extends "admin/change_list.html" %} {% block object-tools-items %}
-<li class="import_resources">
-    <a href="import_resources/">
-        <span style="vertical-align: middle; display: inline-block">
-            Import Resources
-        </span>
-        <span style="vertical-align: middle; display: inline-block">
-            <svg
-                xmlns="http://www.w3.org/2000/svg"
-                width="13"
-                height="13"
-                viewBox="0 0 24 24"
-                style="fill: white; transform: ; msfilter: "
-            >
-                <path d="M11 15h2V9h3l-4-5-4 5h3z"></path>
-                <path
-                    d="M20 18H4v-7H2v7c0 1.103.897 2 2 2h16c1.103 0 2-.897 2-2v-7h-2v7z"
-                ></path>
-            </svg>
-        </span>
-    </a>
-</li>
-{{ block.super }} {% endblock %}
